Added a URL download command line argument.
It is now possible to specify the "-u" argument and a list of URL's (http is currently the only supported URL type at present). This feature depends upon wget being installed on the system. With just the "-u" argument only flat .pkg files are supported. In conjunction with the -i argument you may also specify a .dmg file which may contain .pkg or .mpkg install packages.